use primarySpan.message   consistently in both SemanticsVisitor and CLikeSourceEmitter - Remove stale comment about noteExplicitConversionPossible in   slang-diagnostic-defs.h (file no longer exists) - Change (int) casts to (int64_t) in diagnostic designated initializers   since Lua 'int' type maps to C++ int64_t, avoiding unnecessary   narrowing followed by widening@Address PR review findings for rich diagnostics migration
- Fix inconsistent diagnoseOnce key construction

+++ source/slang/slang-check-conversion.cpp
@@ -997,8 +997,8 @@ bool SemanticsVisitor::_coerceInitializerList(
         if (outToExpr)
         {
             getSink()->diagnose(Diagnostics::TooManyInitializers{
-                .expected = (int)argIndex,
-                .got = (int)argCount,
+                .expected = (int64_t)argIndex,
+                .got = (int64_t)argCount,
                 .initList = fromInitializerListExpr});
         }
     }
